+++ main.tf
@@ -18,6 +18,25 @@ locals {
       }
     ]
   ]))
+
+  generated_map_bucket_obj_admin = distinct(flatten([
+    for bucket in var.buckets_list : [
+      for bucket_obj_adm in bucket.bucket_obj_adm : {
+        bucket_name    = local.generated_bucket_names[bucket.name]
+        bucket_obj_adm = bucket_obj_adm
+      }
+    ]
+  ]))
+
+  generated_map_bucket_obj_vwr = distinct(flatten([
+    for bucket in var.buckets_list : [
+      for bucket_obj_vwr in bucket.bucket_obj_vwr : {
+        bucket_name    = local.generated_bucket_names[bucket.name]
+        bucket_obj_vwr = bucket_obj_vwr
+      }
+    ]
+  ]))
+
 }
 
 # Add a random resource to randomize resource's names to prevent collisions.
@@ -106,6 +125,22 @@ resource "google_storage_bucket_iam_member" "viewer" {
   bucket   = google_storage_bucket.application[each.value.name].name
   role     = "roles/storage.legacyObjectReader"
   member   = "allUsers"
+}
+
+# Default Storage Admin Role
+resource "google_storage_bucket_iam_member" "default_storage_admin" {
+  for_each = { for bucket in local.generated_map_bucket_obj_admin : "${bucket.bucket_name}--${bucket.bucket_obj_admin}" => bucket }
+  bucket   = google_storage_bucket.application[each.value.name].name
+  role     = "roles/storage.objectAdmin"
+  member   = each.value.bucket_obj_admin
+}
+
+# Default Storage Viewer Role
+resource "google_storage_bucket_iam_member" "default_storage_viewer" {
+  for_each = { for bucket in local.generated_map_bucket_obj_vwr : "${bucket.bucket_name}--${bucket.bucket_obj_vwr}" => bucket }
+  bucket   = google_storage_bucket.application[each.value.name].name
+  role     = "roles/storage.objectViewer"
+  member   = each.value.bucket_obj_vwr
 }
 
 # ----------------------